+++ src/main/java/co/develhope/librarymanagement/repository/inventory/WarehouseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WarehouseRepository extends JpaRepository<Warehouse,Integer> {
+public interface WarehouseRepository extends JpaRepository<Warehouse,Long> {
 }
