+++ src/main/java/co/develhope/librarymanagement/repository/library/CustomerRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.time.LocalDate;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer,Integer> {
+public interface CustomerRepository extends JpaRepository<Customer,Long> {
     long countByAbsences(LocalDate absences);
 }
